requirements.txt
pandas
streamlit
plotly.express
plotly.graph_objects